# api/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Stock Management API
  version: 1.0.0
  summary: Stock management for an automotive spare parts shop
  description: |
    Stock management REST API.

    This hand-maintained design document is incomplete and retains historical
    schemas. Check the controllers and DTOs in `stock-infrastructure` for the
    current request and response contracts.

    ## Conventions
    - **Versioning**: URL prefix `/api/v1`; incompatible changes use `/api/v2`.
    - **Case**: `kebab-case` URL segments and `camelCase` JSON fields.
    - **IDs**: UUID v4 strings in all payloads.
    - **Dates**: ISO 8601 UTC `date-time` (`2026-04-23T10:15:30Z`).
    - **Money**: `Money { amount: string (BigDecimal), currency: ISO-4217 }`.
      Amounts use strings to avoid JavaScript precision loss.
    - **Errors**: RFC 7807 (`application/problem+json`).
    - **Idempotency**: POST sales, stock receipts and stock transfers accept an
      optional `Idempotency-Key` UUID. See `IdempotencyFilter` for replay behavior
      and current limitations; no 24-hour expiry is enforced.
    - **Pagination**: `?page=0&size=20&sort=field,asc`, with a `Page<T>` response.
    - **Auth**: JWT in the HttpOnly `access_token` cookie. See `SecurityConfig`
      for public routes and role restrictions; Authorization Bearer is not read by the current filter.

  contact:
    name: Ali Cheikh — ALMA / Nantes Université
    email: maintainer@example.test
  license:
    name: Proprietary
servers:
...
security:
  - cookieAuth: [ ]
paths:
...
  /debts:
    get:
      tags: [ Créances ]
      operationId: listDebts
      summary: Outstanding or settled shop debts.
      description: |
        By default, return outstanding debts, oldest first.
        With `status=SETTLED`, return settled debts, most recently settled first.
        Settled debts remain available as payment records.

        Sorting parameters are not accepted; ordering depends on the debt status.

      parameters:
        - $ref: '#/components/parameters/DebtStatusParam'
        - name: customerId
          in: query
          description: Filter by customer. When omitted, include the whole shop.
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfDebt'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /debts/{saleId}:
    get:
      tags: [ Créances ]
      operationId: getCreditSaleDetail
      summary: Debt details, sale lines, seller and payments.
      description: |
        Return sale lines and quantities, prices at the time of sale, seller,
        customer, and the full payment history with collector names.
        Details remain available after full settlement.

        Return `404` for cash sales without a customer, as well as missing sales.

      parameters:
        - name: saleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditSaleDetail'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token set by the login endpoint.
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (zero-based).
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    DebtStatusParam:
      name: status
      in: query
      description: |
        Filter by settlement status. Defaults to `OUTSTANDING`.
        Request settled debt history explicitly.

      schema:
        $ref: '#/components/schemas/DebtStatus'
...
  responses:
    BadRequest:
      description: Malformed request (syntax validation).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but missing the required role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Money:
      type: object
      required: [ amount, currency ]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d{1,4})?$'
          description: Decimal amount serialized as a string (BigDecimal).
          examples: [ "45.90", "1299.00" ]
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: Code ISO-4217.
          examples: [ EUR, USD ]
...
    DebtStatus:
      type: string
      default: OUTSTANDING
      enum: [ OUTSTANDING, SETTLED, ALL ]
      description: |
        `OUTSTANDING`: a balance remains due.
        `SETTLED`: the debt is paid and remains available for reference.
        `ALL`: both outstanding and settled debts.
    Debt:
      type: object
      required:
        [ saleId, occurredAt, customerId, customerGivenName, totalAmount,
          amountPaid, amountDue, settled, daysOutstanding, overdue ]
      properties:
        saleId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        customerId: { type: string, format: uuid }
        customerGivenName: { type: string }
        customerFatherName: { type: string, nullable: true }
        customerPhoneNumber: { type: string, nullable: true }
        totalAmount: { $ref: '#/components/schemas/Money' }
        amountPaid:
          allOf: [ { $ref: '#/components/schemas/Money' } ]
          description: Sum of payments, derived from the ledger and never stored.
        amountDue:
          allOf: [ { $ref: '#/components/schemas/Money' } ]
          description: Total minus payments received. Zero for a settled debt.
        settled:
          type: boolean
          description: True when no balance remains due.
        settledAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Settlement date, taken from the last payment.
            `null` while the debt is outstanding.

        daysOutstanding:
          type: integer
          minimum: 0
          description: |
            Days from the sale to today for outstanding debts, or to the settlement
            date for settled debts. The count stops at settlement.

        overdue:
          type: boolean
          description: |
            Whether `daysOutstanding` exceeds the 30-day allowance.
            For outstanding debts, indicates overdue payment; for settled debts,
            indicates late settlement.
    CreditSaleLine:
      type: object
      required: [ productId, productName, quantity, unitPrice, lineTotal ]
      properties:
        productId: { type: string, format: uuid }
        productName: { type: string }
        productReference: { type: string, nullable: true }
        quantity: { type: integer, minimum: 1 }
        unitPrice:
          allOf: [ { $ref: '#/components/schemas/Money' } ]
          description: Price at the time of sale, not the current price.
        lineTotal: { $ref: '#/components/schemas/Money' }
    CreditSalePayment:
      type: object
      required: [ paymentId, amount, receivedAt, receivedById ]
      properties:
        paymentId: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
        receivedAt: { type: string, format: date-time }
        receivedById: { type: string, format: uuid }
        receivedByName:
          type: string
          nullable: true
          description: '`null` if the payment collector account no longer exists.'
    CreditSaleDetail:
      type: object
      required:
        [ saleId, occurredAt, sellerId, customerId, customerGivenName, lines,
          totalAmount, amountPaid, amountDue, settled, payments ]
      properties:
        saleId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        sellerId: { type: string, format: uuid }
        sellerName: { type: string, nullable: true }
        customerId: { type: string, format: uuid }
        customerGivenName: { type: string }
        customerFatherName: { type: string, nullable: true }
        customerPhoneNumber: { type: string, nullable: true }
        lines:
          type: array
          description: In entry order, matching the customer receipt.
          items: { $ref: '#/components/schemas/CreditSaleLine' }
        totalAmount: { $ref: '#/components/schemas/Money' }
        amountPaid: { $ref: '#/components/schemas/Money' }
        amountDue: { $ref: '#/components/schemas/Money' }
        settled: { type: boolean }
        payments:
          type: array
          description: |
            Oldest first. The initial deposit, if any, is the first payment.

          items: { $ref: '#/components/schemas/CreditSalePayment' }

    # ── Generic pagination ──────────────────────────────────
    PageMeta:
      type: object
      required: [ page, size, totalElements, totalPages ]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    PageOfDebt:
      type: object
      required: [ content, page ]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Debt' }
        page: { $ref: '#/components/schemas/PageMeta' }

    # ── RFC 7807 Problem Details ──────────────────────────────
    Problem:
      description: |
        RFC 7807 error representation.
        `code`, `errors` and `traceId` are API-specific extensions.

      type: object
      required: [ type, title, status ]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the error type (documentation).
          examples: [ "https://api.stock.example.com/errors/stock-insufficient" ]
        title:
          type: string
          description: Short, human-readable summary.
          examples: [ "Stock insuffisant" ]
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: Detailed explanation of this specific occurrence.
        instance:
          type: string
          format: uri
          description: URI of the affected resource instance.
        code:
          type: string
          description: Stable machine-readable code for frontend mapping.
          examples:
            - PRODUCT_NOT_FOUND
            - STOCK_INSUFFICIENT
            - LOCATION_NOT_FOUND
            - TRANSFER_CROSS_SHOP
            - INVALID_MOVEMENT
            - VALIDATION_FAILED
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [ field, message ]
            properties:
              field: { type: string }
              message: { type: string }
              rejectedValue: { }
        traceId:
          type: string
          description: Correlation ID for locating logs.